# docs/swagger.yaml
basePath: /api/v1
definitions:
  books.BookCreate:
    properties:
      dt_alt:
        type: string
      dt_aqu:
        type: string
      dt_edi:
        type: string
      dt_inc:
        type: string
      nm_assu:
        type: string
      nm_autor:
        type: string
      nm_edit:
        type: string
      nm_isbn:
        type: string
      nm_obra:
        type: string
      nr_edi:
        type: integer
      nr_pags:
        type: integer
      nr_vol:
        type: integer
      sn_ativo:
        type: string
      txt_obs:
        type: string
      usu_alt:
        type: string
      usu_inc:
        type: string
      vr_aqu:
        type: number
    type: object
  books.BookSearch:
    properties:
      nm_search:
        type: string
      nr_docs:
        type: integer
    type: object
  books.BookUpdate:
    properties:
      dt_alt:
        type: string
      dt_aqu:
        type: string
      dt_edi:
        type: string
      nm_assu:
        type: string
      nm_autor:
        type: string
      nm_edit:
        type: string
      nm_isbn:
        type: string
      nm_obra:
        type: string
      nr_edi:
        type: integer
      nr_pags:
        type: integer
      nr_vol:
        type: integer
      sn_ativo:
        type: string
      txt_obs:
        type: string
      usu_alt:
        type: string
      vr_aqu:
        type: number
    type: object
...
  msresponse.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/msresponse.ErrorBody'
      message:
        example: Sucesso
        type: string
      ok:
        example: true
        type: boolean
      request_id:
        example: 7f8c2a9b-5f8a-4b0e-91e5-3b9c2d9a1234
        type: string
    type: object
  msresponse.ErrorBody:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/msresponse.ErrorCode'
        example: 1
      description:
        example: Campo nm_obra é obrigatório
        type: string
    type: object
  msresponse.ErrorCode:
    enum:
    - 1
    - 2
    - 3
    - 4
    - 5
    - 500
    type: integer
    x-enum-varnames:
    - ErrorFormatoInvalido
    - ErrorTokenInvalido
    - ErrorNaoAutorizado
    - ErrorNaoEncontrado
    - ErrorValidacao
    - ErrorInterno
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://swagger.io
  description: This is a template server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io
  title: Server API
  version: "1.0"
paths:
...
  /tabelas/books:
    post:
      consumes:
      - application/json
      description: Insere um novo registro de livro na base de dados. Os campos nome
        da obra e autor são obrigatórios.
      parameters:
      - description: Dados do livro a ser cadastrado
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/books.BookCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Livro cadastrado com sucesso
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "400":
          description: JSON inválido ou campos obrigatórios ausentes
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "500":
          description: Erro interno ao cadastrar o livro
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
      summary: Cadastra um novo livro
      tags:
      - books
  /tabelas/books/{id}:
    delete:
      consumes:
      - application/json
      description: Exclui um livro da base de dados a partir do seu identificador
        MongoDB ObjectID.
      parameters:
      - description: ID do livro no formato ObjectID hexadecimal
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Livro removido com sucesso
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "400":
          description: ID não informado ou inválido
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "500":
          description: Erro interno ao remover o livro
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
      summary: Remove um livro por ID
      tags:
      - books
    get:
      consumes:
      - application/json
      description: Retorna os dados de um livro a partir do seu identificador MongoDB
        ObjectID.
      parameters:
      - description: ID do livro no formato ObjectID hexadecimal
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Livro localizado com sucesso
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "400":
          description: ID não informado ou inválido
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "500":
          description: Erro interno ao buscar o livro
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
      summary: Busca um livro por ID
      tags:
      - books
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um livro existente a partir do seu identificador
        MongoDB ObjectID.
      parameters:
      - description: ID do livro no formato ObjectID hexadecimal
        in: path
        name: id
        required: true
        type: string
      - description: Dados do livro a serem atualizados
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/books.BookUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Livro atualizado com sucesso
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "400":
          description: ID inválido ou JSON em formato incorreto
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "500":
          description: Erro interno ao atualizar o livro
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
      summary: Atualiza um livro por ID
      tags:
      - books
  /tabelas/books/search/nm-obra:
    post:
      consumes:
      - application/json
      description: Realiza busca aproximada de livros pelo nome da obra. Caso o número
        máximo de documentos não seja informado, será utilizado o limite padrão de
        10 registros.
      parameters:
      - description: Parâmetros da busca aproximada
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/books.BookSearch'
      produces:
      - application/json
      responses:
        "200":
          description: Busca realizada com sucesso
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "400":
          description: JSON inválido ou expressão de busca não informada
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
        "500":
          description: Erro interno ao buscar livros
          schema:
            $ref: '#/definitions/msresponse.APIResponse'
      summary: Busca livros pelo nome da obra
      tags:
      - books
...
swagger: "2.0"